requirements.txt
pygithub
furl
